refactor: refactor too many props and large component smell by create ads, docs button, github button, question button, menu, editor change, title and title bar components in layout folder
