+++ src/components/layout/index.tsx
@@ -1,3 +1,3 @@
-export { default as Title } from './Title';
+export { default as Title } from './TitleBar';
 
 export { default as Content } from './Content';
